added kamek fix to std::to_string overload

+++ HashMap.h
@@ -94,11 +94,11 @@ class HashMap {
                     
                     int col=collisions[i][j];
                     if(col>=200&&col<1000) {
-                        string out=" "+std::to_string(col)+ " : "+files[j]+", "+files[i];
+                        string out=" "+std::to_string(static_cast<long long>(col)+ " : "+files[j]+", "+files[i];
                         outputs.push_back(out);
                     }
                     else if (col>=200) {
-                        string out=std::to_string(col)+ " : "+files[j]+", "+files[i];
+                        string out=std::to_string(static_cast<long long>(col))+ " : "+files[j]+", "+files[i];
                         outputs.push_back(out);
                     }
                 }
